backend/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.13.5
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.3
comtypes==1.4.11
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
elevenlabs==2.19.0
eventlet==0.40.3
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.122.1
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==4.1.1
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.10.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-genai==1.55.0
google-pasta==0.2.0
gradio==6.0.1
gradio_client==2.0.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
idna==3.10
imageai==3.0.3
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.15.0
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
limits==5.6.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.2.0
multidict==6.6.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
openai==1.98.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
ordered-set==4.1.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
playsound==1.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.2.0
pyparsing==3.2.3
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.13.0
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safehttpx==0.1.7
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.42
srt==3.5.3
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.8.0
typer==0.20.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uri-template==1.3.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.38.0
vosk==0.3.45
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
wsproto==1.2.0
yarl==1.20.1
